Move buildMetadata to utils.
We now have a clearly-named main script, and the functionality
of building the metadata is in src/utils/buildMetadata.ts.

The code seems more readable to me and easy to follow, and in
the future I expect we break the big buildMetadata() function
into smaller pieces.

Also, the dependency on process.env is removed for buildMetadata,
and the package.json specification is updated, including bumping
version 0.0.1 -> 0.1.0.