Do not add empty fields, and check that there are no missing names.
If a spreadsheet has an empty field, instead of adding it to the
created item as

  item.field = ""

we now do not create the field at all.

This seems to make more sense semantically, and will likely result
in easier future processing of the unspecified fields (especially
when we want to put a category as the default, or leave out some
of the fields because they make sense on only some type of programs
and not others, for example).

Also, if the spreadhseet has a row with data but no name specified,
instead of processing it give an error.